Allow users to opt-in to installing cilium
Since the recommended way of installing cilium is by helm chart and installing a CNI plugin in general is a separate concern, the manifests are included as a convenience measure. Users will be expected to explicitly enable cilium installation by setting the pillar k8s_salt:addons:cilium:enabled:True.